api/requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml[html_clean]==5.3.0
readability-lxml==0.8.1
pdfminer.six==20240706
azure-functions==1.21.3
pandas